Local DNS to help Android emulator resolve oauth2 Docker service running on host machine
Description:
In the Android emulator, "localhost" unavoidably means 127.0.0.1, i.e. the Android host. We want to point to a server running on the host machine, at address 10.0.2.2.

Change the server configuration to use localtest.me instead, which is a publicly registered DNS entry that we can override within the emulator to point to the desired address.

Test Plan:
It should continue to work on the host (DNS not overridden) as well as begin working in the emulator (using the local DNS server described in the android-dns/ subdirectory).

Visit http://oauth2-proxy.localtest.me in a browser in both environments