Android app makes requests to API with Authorization header
Description:
Configures oauth2-proxy to check the header (via confusingly named "skip" param which really means to look at it).

Adds functionality in the app to send network requests to endpoint defined in previous commit.

Test Plan:
Use buttons on logged in screen to send sample requests with and without the header